+++ tests/vm/systemd-multi.nix
@@ -1,8 +1,10 @@
-import <nixpkgs/nixos/tests/make-test-python.nix> {
+{ packages, ... }:
+
+{
   name = "ip2unix-systemd-multi";
 
   machine = { pkgs, lib, ... }: let
-    ip2unix = import ../.. { inherit pkgs lib; };
+    inherit (packages.${pkgs.system}) ip2unix;
 
     testServer = pkgs.writeScript "test-server.py" ''
       #!${pkgs.python3.interpreter}